+++ conftest.py
@@ -1,5 +1,4 @@
 import os
-from copy import deepcopy
 import http
 
 import requests
@@ -63,14 +62,6 @@ def search_results(retrieve_profile_data):
 
 
 @pytest.fixture
-def api_response_200():
-    response = requests.Response()
-    response.status_code = http.client.OK
-    response.json = lambda: deepcopy({})
-    return response
-
-
-@pytest.fixture
 def api_response_search_200(api_response_200, search_results):
     api_response_200.json = lambda: search_results
     return api_response_200
@@ -103,3 +94,42 @@ def feature_flags(settings):
     # solves this issue: https://github.com/pytest-dev/pytest-django/issues/601
     settings.FEATURE_FLAGS = {**settings.FEATURE_FLAGS}
     yield settings.FEATURE_FLAGS
+
+
+@pytest.fixture
+def api_response_200():
+    response = requests.Response()
+    response.status_code = http.client.OK
+    return response
+
+
+@pytest.fixture
+def api_response_404(*args, **kwargs):
+    response = requests.Response()
+    response.status_code = http.client.NOT_FOUND
+    return response
+
+
+@pytest.fixture
+def api_response_search_description_highlight_200(
+    api_response_200, search_results
+):
+    search_results['hits']['hits'][0]['highlight'] = {
+        'description': [
+            '<em>wolf</em> in sheep clothing description',
+            'to the max <em>wolf</em>.'
+        ]
+    }
+    api_response_200.json = lambda: search_results
+    return api_response_200
+
+
+@pytest.fixture
+def api_response_search_summary_highlight_200(
+    api_response_200, search_results
+):
+    search_results['hits']['hits'][0]['highlight'] = {
+        'summary': ['<em>wolf</em> in sheep clothing summary.']
+    }
+    api_response_200.json = lambda: search_results
+    return api_response_200